+++ src/QuickMarcEditor/QuickMarcEditorContainer.test.js
@@ -24,8 +24,6 @@ import buildStripes from '../../test/jest/__mock__/stripesCore.mock';
 import { useAuthorityLinksCount } from '../queries';
 import { applyCentralTenantInHeaders } from './utils';
 
-const mockFetchLinksCount = jest.fn().mockResolvedValue();
-
 const match = {
   path: '/marc-authorities/quick-marc/edit-authority/:externalId',
   url: '/marc-authorities/quick-marc/edit-authority/external-id',
@@ -156,10 +154,6 @@ describe('Given Quick Marc Editor Container', () => {
 
   describe('when the marc type is authority', () => {
     it('should make a request to get the number of links', async () => {
-      useAuthorityLinksCount.mockClear().mockReturnValue({
-        fetchLinksCount: mockFetchLinksCount,
-      });
-
       await act(async () => {
         await renderQuickMarcEditorContainer({
           mutator,
@@ -170,7 +164,7 @@ describe('Given Quick Marc Editor Container', () => {
         });
       });
 
-      expect(mockFetchLinksCount).toHaveBeenCalledWith([match.params.externalId]);
+      expect(useAuthorityLinksCount).toHaveBeenCalledWith({ id: match.params.externalId });
     });
   });