+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-quick-marc
 
 
+
 ## [8.0.0] (IN PROGRESS)
 
 * [UIQM-580](https://issues.folio.org/browse/UIQM-580) *BREAKING* Move QuickMarcView, PrintPopup and getHeaders to stripes-marc-components.